+++ packages/frontend/src/pages/dashboard/index.tsx
@@ -4,12 +4,11 @@ import Head from "next/head";
 import { useGetUserGuildsQuery } from "../../generated/graphql";
 import LoginButton from "../../components/LoginButton";
 import GuildListComponent from "../../components/Dashboard/GuildListComponent";
-import { withUrql } from "../../utils/withUrql";
 
 interface DashboardPageProps {}
 
 const DashboardPage: NextPage<DashboardPageProps> = () => {
-  const guildQuery = useGetUserGuildsQuery()[0];
+  const [guildQuery] = useGetUserGuildsQuery();
 
   return (
     <NavbarProvider>
@@ -26,7 +25,7 @@ const DashboardPage: NextPage<DashboardPageProps> = () => {
       ) : guildQuery.error ? (
         <div className="flex flex-col items-center gap-y-2">
           <p>{guildQuery.error.message}</p>
-          <LoginButton className="bg-gray-800 py-2 px-5 rounded" />
+          <LoginButton className="bg-slate-800 py-2 px-5 rounded" />
         </div>
       ) : (
         <>
@@ -39,4 +38,4 @@ const DashboardPage: NextPage<DashboardPageProps> = () => {
   );
 };
 
-export default withUrql(DashboardPage);
+export default DashboardPage;